refactor: splite quiz Practic section Task in technicalInterview component to smaller and extract custom hooks
