+++ client/src/components/Dashboard/LearnSection/Recruitment/components/TechnicalInterview/PracticeSection.component.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { motion } from 'framer-motion';
 import { FaArrowLeft } from 'react-icons/fa';
 import { usePracticeSection } from './components/Practice/hooks/usePracticeSection.hook';
-import TaskDetails from './components/Practice/TaskDetails.component';
-import TaskCard from './components/Practice/TaskCard.component';
+import TaskDetails from './components/Practice/Task/components/TaskDetails.component';
+import TaskCard from './components/Practice/Task/components/TaskCard.component';
 import StatsSection from './components/Practice/StatsSection.component';
 import FilterSection from './components/Practice/FilterSection.component';
 import EmptyState from './components/Practice/EmptyState.component';
